requirements.txt
PyYaml==3.12
jinja2
glob2
python-keystoneclient
python-neutronclient
python-novaclient==3.3.3
requests

